check IsActive exists before dropping it

+++ database/migrations/2021_06_01_135739_add_is_active_column_to_governments_table.php
@@ -25,8 +25,10 @@ class AddIsActiveColumnToGovernmentsTable extends Migration
      */
     public function down()
     {
-        Schema::table('governments', function (Blueprint $table) {
-            $table->dropColumn('IsActive');
-        });
+        if (Schema::hasColumn('governments', 'IsActive')) {
+            Schema::table('governments', function (Blueprint $table) {
+                $table->dropColumn('IsActive');
+            });
+        }
     }
 }
